requirements.txt
pytest

# 0.2
fastparquet
pandas
pyarrow

# 3.
scipy

# 4.
scikit-learn

# 6.
annoy
tensorflow-hub
tensorflow-macos

# 7.
pystemmer
regex
uniseg

# 8., 9.
requests
